mini/llvm: clear the translator's dead code and enable -Wswitch
Cleanups left deliberately in place by the mechanical six-way split of
translator.cpp, now that the move itself is settled.

Deleted as dead:
  - LLVMFunctionType3/4/5.  All three are static in translator-emit.cpp,
    so only that TU could reach them, and it has no callers.  The
    apparent callers in mini-llvm.c (:3173, :4510) resolve to that file's
    own private copies at :1642/1658/1676.  LLVMFunctionType0 is
    non-static and used from translator-bb.cpp; 1 and 2 have in-file
    callers; all three stay.
  - The tmp_ji MonoJumpInfo built by get_aotconst (), whose two assigned
    fields are never read.
  - Set-but-unused stack_size, and `name` in init_jit_module ().

Deleting `name` also removes a genuine per-domain leak: the
g_strdup_printf ("mono-%s", domain->friendly_name) result was assigned,
never read and never freed.  Nothing observed it -- the module is created
with the hardcoded name "jit-global-module" -- but dropping a leak is a
behaviour change, small as it is.

Fixed the -Wsign-compare sites.  bb_index becomes guint: both loops
ascend from 0 against unsigned bounds (cfg->num_bblocks and
bblock_list->len), there is no descending loop or >= 0 termination, and
no %d consumer.  The six `bb->region != -1` comparisons now all read
`bb->region != (guint)-1`, which is value-preserving by the standard,
rather than narrowing region to int; previously only two sites were
touched and the tree carried two spellings of one idiom.

For MAX (..., cfg->method_index): mini.c:3279 assigns method_index only
under if (cfg->compile_aot), and the sole AOT caller
(aot-compiler.c:8876) passes a real table index in [0, num_methods).
Every JIT caller passes -1 but never reaches the assignment, so the field
stays 0 from the MonoCompile allocation and the cast is value-preserving
for all reachable values.  Worth recording the hazard this hides: had a
JIT path ever set method_index = -1, the original mixed-signedness
comparison would have promoted max_method_idx to -1, where the cast makes
it 0.

Enabled -Wswitch for mono/mini's C++ by adding it to AM_CXXFLAGS.  The
-Wno-switch it counteracts is not a mono/mini flag at all: configure puts
it in the global CPPFLAGS, whence it reaches every TU in the tree, so it
cannot be dropped per-directory.  All 58 warnings came from two switches
-- the WrapperInfo subtype switch in translator.cpp (28) and the inner
ptr_type switch in translator-types.cpp (30) -- and both are
"handle a few, ignore the rest" over large enums, so both get a
default: break; rather than ~50 enumerated no-op cases.  Verified
behaviour-neutral: control already fell out of each switch to the
following statement.  --enable-werror cannot turn this into a build
break; SHARED_CFLAGS, which carries -Werror, is referenced only from
AM_CFLAGS.

MonoLLVMModule::used stays.  Its writer, the OP_OBJC_GET_SELECTOR case in
process_bb (), is compiled unconditionally -- it is not inside
#ifdef TARGET_OSX -- so the field is needed on every target; the comment
claiming otherwise is corrected.  That the field is write-only, and its
GPtrArray never freed, is left for its own change.

The 13/13 unit-test result is not evidence for any of this: those tests
cover the ORC engine, relocation widths and .eh_frame, and none of them
calls mono_llvm_emit_method.  The coverage for these changes is the
regression runs under MONO_TIERED=1 --llvm, where
mini_tiered_promote -> mini_method_compile -> mono_llvm_emit_method is
confirmed reached under gdb: the full exceptions.exe matrix (2210 tests
x 26 opt combinations) plus an mcs.exe self-compile.  A -Wswitch rebuild
of every C++ TU in mono/mini emits zero warnings.

+++ mono/mini/llvm/translator-bb.cpp
@@ -541,7 +541,7 @@ process_bb (EmitContext *ctx, MonoBasicBlock *bb)
 				emit_volatile_store (ctx, ins->next->dreg);
 			} else if (MONO_IS_COND_EXC (ins->next)) {
 				gboolean force_explicit_branch = FALSE;
-				if (bb->region != -1) {
+				if (bb->region != (guint)-1) {
 					/* Don't tag null check branches in exception-handling
 					 * regions with `make.implicit`.
 					 */
@@ -4263,7 +4263,7 @@ process_bb (EmitContext *ctx, MonoBasicBlock *bb)
 			 */
 		case OP_IMPLICIT_EXCEPTION:
 			/* This marks a place where an implicit exception can happen */
-			if (bb->region != -1)
+			if (bb->region != (guint)-1)
 				set_failure (ctx, "implicit-exception");
 			break;
 		case OP_THROW:
